App_config created + Nlog is included
