+++ .hg/hgrc
@@ -1,0 +1,1 @@
+# Generated by TortoiseHg
