+++ .hg/requires
@@ -1,0 +1,6 @@
+dotencode
+fncache
+generaldelta
+revlogv1
+sparserevlog
+store
