+++ OrderCreator.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28803.452
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OrderCreator", "OrderCreator\OrderCreator.csproj", "{71A641B8-601D-4274-BBEA-801BB008AEB3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OrderCreator", "OrderCreator\OrderCreator.csproj", "{71A641B8-601D-4274-BBEA-801BB008AEB3}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{20F21F0C-623C-4926-B934-E21DB0ACE66A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
