+++ OrderCreator/.hg/hgrc
@@ -1,0 +1,11 @@
+# Generated by TortoiseHg
+
+
+[web]
+name = OrderCreator
+
+
+[http_proxy]
+host = https://github.com/yarkablukov/OrderCreator
+user = yarkablukov
+passwd = GithubTransaction0110
