+++ OrderCreator/App.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <appSettings>
+    <add key="customer_orders_directory" 
+         value="C:\Users\Slava\source\repos\OrderCreator\OrderCreator\bin\Debug\netcoreapp2.1\customer_orders" />
+    <add key="delivery_orders_directory"
+            value="C:\Users\Slava\source\repos\OrderCreator\OrderCreator\bin\Debug\netcoreapp2.1\delivery_orders" />
+    <add key="xsdSchemaPath"
+                value="C:\Users\Slava\source\repos\OrderCreator\CutomerOrderXMLSchema.xsd" />
+  </appSettings>
+</configuration>
