+++ OrderCreator/Orders/OrderGoods.cs
@@ -4,11 +4,5 @@
 
 namespace OrderCreator
 {
-    public class OrderGoods
-    {
-        public string name;
-        public float price;
-        public int quantity;
-        public float sum;
-    }
+
 }
