+++ OrderCreator/Services/Logger.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using log4net;
+using log4net.Config;
+
+namespace OrderCreator.Services
+{
+    public static class Logger
+    {
+
+    }
+}
